Frontend setup: Add settings for recording quality
Recordings are marked as DAMAGED based on recording quality.
Add settings that let the user customize the quality calculation
by specifying how much missing time is allowed at the start and end of
recordings and the quality required for not marking as DAMAGED.
Previously the times were hardcoded as 15 seconds and the
percentage as 95.