clazy: Fix 'allocating an unneeded temporary container' warning.
Don't create a list of all the keys in the map, select one, and then
index back into the map. With a large map and a small index this
wastes time in creating the list of keys. Just run an iterator over
the map and move the right number of steps. This code works in both
Qt5 and Qt6. (Qt5 let you add an integer to an iterator to move that
many items. Qt6 only allows you to increment the iterator.)