Finish cleaning up bad mapping values.
Fix the mappings that are returned from the ProgramInfo::ToMap,
BlurayMetadata::toMap, RecordingRule::ToMap, and ThemeInfo::ToMap
functions. Several of the values are stored as a single 16-bit QChar,
potentially losing the high order bits of the number/bitfield.

A set of earlier fixes added new keys with properly stringified
values, leaving the old broken values in place. This commit removes
the broken values completely, and uses the original key names for the
properly stringified values.

Fixes #366.