Don't cache iterators across a call to a Qt ::erase function. (mythcommflag)
Don't try to cache the "next" iterator across a call to QHash::erase
or QMap::erase. These functions can internally rearrange the contents
of the hash and invalidate any iterators. They do explicitly return
the next iterator after the one that was erased, so use that value to
continue walking the hash.