Eliminate duplicate calls to TV::customEvent.
The TV::customEvent function receives a number of MythTV generated
events directly from the Qt event dispatcher. The TV object also
registers to filter all events that are about to be sent to the
MythMainWindow object, so it sees these same events a second time in
the TV::filterEvent function before they are allowed to be sent to the
MythMainWindow::event function. TV::eventFilter should not forward
these events to TV::customEvent for processing, since they've already
been processed by that function.