US ATSC uscable channel 159 removed from channel scan
When doing a channel scan with mythtv-setup for ATSC cable, choice "uscable",
the scan range includes channel 159 corresponding to a frequency of 1005MHz.
Scanning this channel with a HDHomeRun fails with an "ERROR: invalid channel" error message.
Consulting the Wikipedia, https://en.wikipedia.org/wiki/Pan-American_television_frequencies,
shows that the highest channel number is 158 and that channel 159 does not exist.