DB access for information about existing transports
In mythtv-setup, the channels found in a scan are compared with the channels in the database.
Do not retrieve the sourceid from the the first channel of the first transport
anymore but use the sourceid that is passed as parameter.
This fixes a possible segmentation fault when the first transport does not have
any channels at all due to filtering of unwanted channels.

(cherry picked from commit 7f48b7007e387030822874d810c453985e7331b8)
Signed-off-by: Klaas de Waal <kdewaal@mythtv.org>