Fix videodev2.h compilation errors on FreeBSD.
The file v4lchannel.cpp should include our own copy of videodev2.h.
Make the same change to this file as was made to the v4l2 files in
commit 730975bc18.

Note: This change has not been made to the development branch, as the
local copy of videodev2.h has been completely removed there.