ATSC Scan Range menu page
In the ATSC Scan Range menu page it is possible to reduce the
scan range by selecting the first and the last channel to be scanned.
For each of the three items First Channel, Last Channel and Channel Count
a label text and help text have now been added.
Also the Channel Count item is now read-only.
There are no functional changes.