US ATSC uscable channels T7 to T12 removed
When doing a channel scan with mythtv-setup for US ATSC cable, choice "uscable"
(or "Cable" depending on the translation), the scan range includes channels T7 to T14
corresponding to a frequency range from 8.75 to 50.75MHz.
Consulting the Wikipedia, https://en.wikipedia.org/wiki/Pan-American_television_frequencies,
shows that the 'Subband "T" channels' are indeed valid channels for cable TV.
However, scanning these channels with a HDHomeRun fails with an "ERROR: invalid channel"
error message for channels T7 to T12. Therefore the channels T7 to T12 have now been removed.
It can be argued that channels T13 and T14 can also be removed as there are no known
service providers that use these channels. However, as it cannot be ruled out that there
are private cable networks that use these channels and because the HDHomeRun is
capable of receiving these channels they remain.
Channels T13 and T14 have been moved to the start of the scan so that the whole cable scan
is done monotonous from the lowest frequency to the highest frequency.