MythFrontend: Fix valgrind error in ExitPrompter
In the destructor of the ExitPrompter of mythfrontend the MythDialogBox is
reconfigured with a SetReturnEvent call so that it does not respond to
keypresses anymore after the ExitPrompter is deleted.
According to the description in commit 9f1005638a34e2a2b20632e09fe093adfddbe1b,
where this functionality was introduced, this is a fix for a crash which can
happen if the ExitPrompter is deleted before the MythDialogBox.
The problem with this fix is that it, as reported by valgrind, accesses the
MythDialogBox via m_dialog after the MythDialogBox is deleted.
The delete of MythDialogBox does not clear m_dialog.
The delete of MythDialogBox is triggered by qApp->exit() in ExitPrompter::DoQuit.
Fixed by moving the SetReturnEvent call to DoQuit, before qApp->exit() is called.