Information pages for conflicting recordings
In the program guide of mythfrontend, pressing "R" does schedule a recording
of the program that is currently highlighted. If it is not possible to schedule
the recording the status message "Single Record 0 Conflicting" is shown.
In theme MythCenter-wide this message appears somewhere in the right-top corner.

Pressing "R" again now gives an information page with the title and subtitle
of the selected program, followed by some text and then a list of the programs
that are recorded instead. Each conflicting program is presented with starttime,
endtime, title and subtitle. This allows you to make a decision on which of the
programs you do want to record.

A similar page can be shown as "Preview" for the "Edit Recording Rule" page
to evaluate changes in the "Schedule Options".

The following changes have been made:

- The selected program is now shown also with starttime and endtime.
This makes it easer to see the overlap between the selected program
and the conflicting programs.

- Only conflicting programs from the same video source are shown.
Previously, all programs from all video sources that overlapped in time with
the selected program were shown. This is not correct because a program from
another video source will be recorded using another tuner. Cancelling that
recording will not make any difference for recording the selected program.

- No limit on the maximum number of conflicting programs shown.
The actual number shown does depend on the theme; for MythCenter-wide the limit is 3,
for MythCenter the limit is 4 but Terra can show 8 conflicting programs.
Previously there was a limit of 4 in first described conflicting recordings page
but there was no limit in the "Preview" conflicting recordings page.

- The code is somewhat simplified.
Although also for C++ code the old saying "Beauty is in the eye of the beholder" is
more valid than ever. Valgrind is not angry but other tools might be....