Title for full-screen MythDialogBox
A popup window, name MythPopupBox, does not have a title but a full screen
window, name MythDialogBox, does have the capability to show a title.
The MythDialogBox is used when the dialog is to be shown full screen.
Commit 935863469bbfc36963bd389b07e62883a2cae361 does remove the title for
both variants of the dialog box because the code in MythDialogBox::Create
is used for both.
This commit restores the title but only for the full-screen MythDialogBox.
Note that the default and the MythCenter themes do have a "title" textarea
defined for the MythDialogBox but the Terra theme does not have a "title" textarea.