Services API: Add IgnoreLiveTV/IgnoreDeleted to GetRecordedList
Add optional boolean IgnoreLiveTV and IgnoreDeleted
parameters to Dvr/GetRecordedList.

Error checking added in case RecGroup is set to either
LiveTV or Deleted. In case this is backported to v31,
set the Ignore* parameter to false and log the error.
Better than throwing an error because that could change
the existing response (unlikely, but true).

Also modifies LoadFromRecorded() that does the actual SQL.

Will be manually added to the V2 API later.