Default value for Starting Channel in mythtv-setup
In mythtv-setup page 'Input Connections" the Live TV start channel
can be selected. This value is later stored in the database.
Before storing the value check if it is valid and if not, replace
it with a suitable default value which can be any visible channel
of the video source connected to that capture card.
An invalid starting channel is therefore now no longer mentioned
in the mythtv-setup exit popup dialog.
This supports a workflow where the 'Input Connections'
are defined before the channel scan is done.
Updated the help text of the "Starting Channel" item in mythtv-setup.
Updated the message texts in the mythtv-setup exit popup dialog so
that more messages do fit in the box.