Tweak setting of original airdate in recordings.
If the TV listings don't provide an original airdate, add a step when
computing an original airdate from the provided year.  Instead of
always using Jan 1st of the provided year for the original airdate, if
the year is the current year then use the recording start date for the
original airdate.