HD ringbuffer size spinbox maximum value
Decrease the maximum value as shown in the spinbox from 96256 to 94000.
In mythtv-setup, 1. General, Miscellaneous Settings, File Management
Settings, the HD Ringbuffer Size can be configured.
The spinbox description is "Values from 4700 to 96256 in increments of 4700".
However, the maximum value that actually can be entered is 94000 which is
the highest multiple of 4700 below 96256.