CI/CAM: Connection establishment for not spec compliant CAMs broken
This is for MythTV 0.27.1.

I do have a CAM that unconditionally sends T_SB for tcid 2 on connection open which is not spec compliant.
MythTV basically provisions for such cases in cCiTransportConnection::CreateConnection?().
However, cCiTransportConnection::RecvTPDU() breaks this as it unconditionally calls Init() on
any "error", thus resetting the processing state from CREATION to IDLE.
The attached patch fixes this for my case and should be seen as a sample workaround.
Actually some interworking between CreateConnection?() and RecvTPDU() is required to handle
the retry cases in state CREATION properly.

20210824 Tested on today's master.

Fixes #trac-12184