Automatically extend sports recordings.
This initial implementation checks the game status on the ESPN website
every ten minutes, and will extend recordings if the game isn't
over. Recordings must be flagged to use this behavior, and this sport
must be known to this code. The code is designed to be easy to add
additional data providers for checking on games status.

Closes #409.