qt6: Make tests work again.
When compiling libraries, the qt6 qmake adds "-Wl,-rpath,/usr/lib64"
to the end of the link line for the library. This has the result of
embedding a RUNPATH=/usr/lib64 entry into the dynamic section of the
shared library. This overrides the RPATH in the test executable, and
prevents any of the transitively dependent libraries from being
found. (I.E. test_foo loads libmythmetadata which is found via RPATH,
but libmythmetadata needs libmythbase which can't be found because
RUNPATH overrides RPATH and only specifies /usr/lib64.)

Make the tests work on Qt by telling it not to set any RPATH
variables.