Default tuning parameters for satellite reference transponders
In the mythtv-setup "Full Scan (Tuned)" dialog for satellites, add the
option to select a satellite for setting the initial tuning frequency
and scan parameters such as polarity and symbolrate to start scanning
for channels on that satellite.
There are values for the Hotbird 13 and the Astra 1/2/3 satellites.
Tuning parameters for other satellites can be added.
N.B. The values are hard-coded so they will not remain correct forever.