Do not perform allatonce grab when --only-update-channels specified
Performing an allatonce grab can result in a grabber obtaining
and producing a large amount of data which can take a long time
and/or use a lot of resouces.  Since --only-update-channels is
specified to obtain as little data as possible in order to update
the channels, fall back to a single day grab even for grabbers
that prefer allatonce.