Use grabber --list-channels option if supported for --only-update-channels
Grabbers that support --list-channels will return
just the channel data, and may not use resources to
obtain and/or format program data.  It is a good
fit for an --only-update-channels request when
available.  Some grabbers that do not support
--list-channels may perform equivalently with a
--days 0 grab request, but --days 0 is not in spec,
and is rejected by some grabbers, so we cannot use
that as an alternative.