Cache ttvdbv4 bearer token for the 'ttvdb4.py' grabber
for a timespan configured in 'ttvdb4.ini'.

TheTVDB.com says, that this token is valid for 1 month.

Note: In case of a '--debug' switch supplied to 'ttvdb4.py',
the token cache will be updated. This ensures, that we always
use the latest supplied token.

For now, the lifetime is set to 1 day (24 hours). This ensures,
that a mass update of the metadata likely uses the same token.