Fix "failed to disable screensaver" warning message.
This warning happens because the DBus service used to disable the
screensaver has changed over time. MythTV currently supports four
different DBus services for disabling the screensaver. The existing
code does detect when an attempt to connect to a service
fails. However it doesn't detect when the connection to the service is
successful, but then invoking the service fails.

Add code that immediately after connecting to the service, tests to
see whether or not the screensaver can actually be disabled. If not,
ignore this dbus service.

Fixes #394.