Live TV channel up/down for SPTS inputs
Live TV channel changes for Single Program Transport Stream inputs, such as can be
received from a HDHomeRun with the ExternalRecorder and mythhdhrrecorder, can fail.
This is indicated by a popup error message "Error opening jump program file buffer"
together with a number of repeating error log messages like this:
2021-10-04 21:57:21.472913 E  DTVRec[12]: HandleSingleProgramPAT(NULL)

This is likely to be caused by latencies in the input chain which, after a channel
change, can cause a PAT of the previous channel to be processed as if it were the PAT
of the new channel. When this happens the data of the new channel is filtered out and
thus nothing is received anymore, leading to the aforementioned error messages.
It is likely that this also causes problems with adjacent recordings on different channels.
This is fixed by updating the desired program from the latest PAT whenever the desired
program is not present in the PAT.