Recode SPTS PAT logic in DTV signalmonitor
Recode the logic in the PAT handling for the Single Program Transport Stream.
SPTS without a valid program ID is a supported format used by IPTV and ExternalRecorder.
Error messages like the following:

2021-10-04 21:41:48.522019 E [908876/909057] ExternSH recorders/dtvsignalmonitor.cpp:328 (HandlePAT) - DTVSigMon[12]: Program #1 not found in PAT!
Program Association Section
  PSIP tableID(0x0) length(13) extension(0x807)
       version(0) current(1) section(0) last_section(0)
       tsid(2055) programCount(1)
  program number 19043 has PID 0x0a28

2021-10-04 21:41:48.522021 E [908876/909057] ExternSH recorders/dtvsignalmonitor.cpp:333 (HandlePAT) - DTVSigMon[12]: But there is only one program in the PAT, so we'll just use it

which were given on every channel change have now been replaced by an informational message like this:
2021-10-06 17:02:24.020853 I  DTVSigMon[12]: Program 19043 found in PAT of SPTS

The error message is still given when a program cannot be found in a stream with
multiple programs and in that case the recording fails.