Implement queuing of http requests
With this change rather than immediately processing the
incoming request, it's placed on a queue, and a signal
is emitted to request the work is processed.

If there are available threads available then the incoming
request is immediately processed.

If no threads are available we wait until one of the existing
threads signals its completion, then check to see if there is
anything in the queue to process. If there is, then we trigger
processing of the queued incoming connection.

Fixes: #397