New V2ervice method: Status/GetBackendStatus
This method provides the same info as Status/xml but uses the
standard service model, and is thus available in json as well as xml.
Also some parts of the xml service do not work, both the frontend list
and backend list return nothing. That is fixed in the new
GetBackendStatus method.

Some code has been moved for other service methods to v2serviceUtil.cpp
so that GetBackendStatus can use common code rather than having two
copies of the code.