V2 Services: Ability to prevent inapplicable structures output
Adds the ability to set a property "isNull" to true in an embedded class
to prevent the output of default values where there is no valid value.

Implemented "isNull" on the "Recording" tag of Encoder so that the
Recording details are only shown if there is a recording in progress.