cppcheck: Fix "call of pure virtual function in constructor" warning.
The line this error points to is not wrong because its not an actual
call to the pure virtual function, only setting up a callback for
later.  However, it does point out a potential problem. If someone
instantiates an object of this class and the callback is called,
there is no function there to handle the callback and the program will
crash.

Fix this by marking the constructor as "protected" so its not possible
to instantiate an object of this class. Since all sub-classes must
provide a ColourSpaceUpdate function, it shouldn't be possible to get
into a situation where the callback crashes.  Tag the code to suppress
this warning since its now been handled.