Fix "suspicious call argument" warning in DeviceReadBuffer.
Tidy thinks that the first two arguments of this function might be
swapped because the declaration of the function being called has
"read_len" as its first parameter, and the second parameter to the
function is "read_size". This commit changes the first argument from
"len" to "read_len" to eliminate the warning.