tidy: Fix "unhandled exception at new" warnings in networkcontrol.cpp.
According to the C++ standard, calls to new can throw an exception
upon failure or return a null pointer.  Specifically indicate that
this call may not throw but must return a null pointer, and handle the
failure case. The only result of an allocation failure here is that a
courtesy "shutting down" message isn't sent to clients.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-unhandled-exception-at-new.html