Clean up SPEEDINC/DEC support and INFO playback OSD.
Remove support for the non-slomo speeds.  Those speeds are already
supported more cleanly by either timestretch or fast-forward.  This
effectively makes the SPEEDINC/DEC support a slow motion only
function.

Also, make the OSD, "play" message when INFO is pressed be consistent
with the OSD message that was displayed when the speed was originally
set, especially, fast-forward and rewind speeds.