clazy: Use 4-arg connect in globalsettings.
Clazy recommends using the 4-arg variant of Object::connect over the
3-arg version. The main reason for this is that the signal will not be
called if the target object has been deleted.

In this instance, the lambda doesn't really require an object since is
only calling static functions, and functions on a global object. Use
the sender object as the receiver, which should guarantee that the
receiver object is always valid.

https://github.com/KDE/clazy/blob/master/docs/checks/README-connect-3arg-lambda.md