Remove python oursql connector, not maintained anymore
In regard to oursql, there are still a view packages (1) for python3
and the documentation (2) states compatibility to python3.
The git location has a branch for python3:
https://github.com/python-oursql/oursql/tree/py3k,
which is not up to date (last change 2016).

Once, this project is active again, we can easily re-enable
the oursql connector.

(1) pypi:oursql3 https://pypi.org/project/oursql3/
    arch linux: python-oursql 0.9.5-1, last maintained 2018
    to fix python3.7 compatibility, based on a fork at
    https://github.com/cinatic/python-oursql

(2) https://pythonhosted.org/oursql/

Refs #422