Modernize class OrdDict from MythTV.altdict.py
Dictionaries are ordered since Python 3.6 and PEP 468 (Preserving
the order of **kwargs in a function.) is in place.
This means, that there is no need to service the accompanying list
called '_field_order' in a special way. I is simmply the list(keys()).

Note: This commit follows the python3 habit, that dict.keys(),
dict.values() and dict.items() are view objects:
View objects are dynamic and reflect dict change.
To make these view objects static, one have to convert this view objects
to a list, like 'list(dict.values())'.

Changed view objects to list objects accordingly in the MythTV module.

Legacy implementations for iterkeys(), itervalues(), iteritems() are
provided.

Refs #422