Add experimental procedural texture support
Inspired by https://www.shadertoy.com/ this adds some procedural
texture support (think computer generated art) to the UI code.
Essentially this allows you to run custom OpenGL shaders embedded in
the theme.

The initial use will be for the MythCenterXMAS theme to use as a snow
overlay on top of the background that shows gently falling snow on top
of theme's Christmas snowy background images.

There is a new MythUI type <procedural> that simply defines the area the
shader effect will draw in and the name of two files that contain the
vertex shader program and the fragment shader program.

<procedural name="backimg">
    <area>0,0,100%,100%</area>
    <vertexsource>vertex_shader.txt</vertexsource>
    <fragmentsource>fragment_shader.txt</fragmentsource>
</procedural>

Based on a patch from Mark Kendal.