Only send a single SSDP byebye message.
There are 8 different UPnP byebye messages that get sent
at shutdown of the upnp/ssdp stack.

The UPnP Notification task would repeat the same message
after a delay of up to 250ms.

The repeated message is now skipped on shutdown improving
the shutdown speed.

However the shutdown sequence is called twice. Once from
UPnp::Cleanup() and once in the SSDP destructor. This is
not addressed in this commit.

This improves shutdown time by around 4 seconds in total.

Refs #274