Fix DeprecationWarning: invalid escape sequence
on regex strings

From python 3.7+ docs: (*)
Regular expressions use the backslash character ('\') to indicate
special forms or to allow special characters to be used without
invoking their special meaning. This collides with Python’s usage
of the same character for the same purpose in string literals; for
example, to match a literal backslash, one might have to write '\\\\'
as the pattern string, because the regular expression must be \\,
and each backslash must be expressed as \\ inside a regular Python
string literal. Also, please note that any invalid escape sequences
in Python’s usage of the backslash in string literals now generate a
DeprecationWarning and in the future this will become a SyntaxError.
This behaviour will happen even if it is a valid escape sequence for
a regular expression.

The solution is to use Python’s raw string notation for regular
expression patterns; backslashes are not handled in any special
way in a string literal prefixed with 'r'.
So r"\n" is a two-character string containing '\' and 'n', while "\n"
is a one-character string containing a newline. Usually patterns
will be expressed in Python code using this raw string notation.

Conclusion:
Mark regex strings as 'raw' where necessary.
Note: The 'ru' or 'ur' prefix for strings is not allowed in python3.

(*) https://docs.python.org/dev/library/re.html

Refs #422