Fix error with python3.10 on MythTV/msearch.py
Traceback (most recent call last):
  File "/.../MythTV/msearch.py", line 35, in __init__
    self.sock.setblocking(0.1)
TypeError: 'float' object cannot be interpreted as an integer

Python's socket implementation allows to set the blocking mode with
 - setblocking([0/1] or
 - settimeout(float)  # i.e '0.0'

Prior to python3.10, 'setblocking' evaluated the float value to an integer,
which results in this case to zero (tested with 'socket.gettimeout'()).

Make it clear, that we want the non-blocking mode.

Tested with python3.6 and 3.10.