Fix uninitialized memory access in goom visualiser
The random number table with 65536 ((0x10000) elements was
initialized except for the first element.
The initialization code is changed to a for loop that does
initialize all elements of the table.
Note that the while loop in the original code was only
somewhat correct if the size of the table, given in NB_RAND, was
equal to the maximum value of a 16-bit unsigned integer plus one.

Refs #429