Fix crash of goom visualisation in MythMusic
The value returned by iRAND is used as an array index in
visualisations/goom/tentacle3d.cpp at line 188:
  float val = (float)(ShiftRight(data[0][iRAND(511)],10)) * rapport;
The problem is that iRAND can return negative values and this
is sufficient to cause the problems mentioned in issue #429.
Fixed by changing the iRAND function to return an unsigned integer value.
Note there is a commented-out implementation of iRAND present in the code;
this implementation does also return an unsigned integer value.

Refs #429