Changed "waiting for tsid" message from INFO to DEBUG
Changed the log message about "Received PAT for tsid XX waiting for tsid YY" from INFO to DEBUG.
This is done to avoid polluting the logs when not debugging.
Note that This message is useful for investigating the "Rescan your transports" warning messages
that are sometimes given even when the scanning information is completely up-to-date.