Un-break build of MythNetVision for Debian/Ubuntu
Some background information:
Debian builds their python-only packages ('pure-lib') under
$(builddir)/usr/lib/python3/dist-packages
and installs them under /usr/lib/python3.x/dist-packages.
The packaging scripts from
https://github.com/MythTV/packaging/tree/master/deb
for the MythTV Plugins need to find these intermediate builds to check for
the existance of the MythTV python bindings in advance.

Nevertheless, python3.10 deprecates the 'distutils' package, and python3.12
will remove this package, which provides the correct paths to the python libs.
As of now, there is no alternative module for Debian available, which
provides this information, i.e.: find the path to intermediate python libs.
This implementation priovides a - guessed - mix of 'sys.path' entries,
in the hope that one will fit in the future.

Refs #437