Add percentage finished to "View Recordings" and "Upcoming Recordings".
Buttonlist buttons can now contain a progressbar, as an intuitive way
to show contextual "progress".  This is up to the theme.

"Watch Recordings" uses it to show playback position. Previews aren't
regenerated whilst a file is in use, so playback by another frontend
only updates the progress bar/counter, until playback ceases.

"Upcoming Recordings" uses the bar to show the progress of time for
recordings - it will progress to 100% for failing (0-byte)
recordings. Filesize is also available for "Upcoming Recordings" now
(although Scheduler Updates will reset it to 0.00 for 10 secs)

A percentage progress value is also provided that can be used for
sorting, or as a "depends" for a 'part-watched' icon. This can also be
used in text templates.

The modified MythCenter and MythCenter-Wide themes demonstrate the UI
changes by using putting a progress bar behind the show title in the
"View Recordings" and "Upcoming Recordings" pages.

Video Library has no UI indication of progress.

This is based on Roger Siddons work from 2015. See
https://code.mythtv.org/trac/ticket/12809 for the full details.

---

The MythUIButtonListItem now supports two progress bars called
"buttonprogress1" and "buttonprogress2" that may be used for any
purpose.

The ProgramInfo object now provides new "recordedpercent" and
"watchedpercent" strings, that can be use to display the percentage of
the program already recorded/watched at any place that other items can
be displayed from the result of ProgramInfo::ToMap. These values are
not continuously calculated, but must be set by the frontend code
before calling ProgramInfo::ToMap.

The "watchrecordings" screen described in the "recordings-ui.xml"
theme file supports two additional progressbars ("recordedprogressbar"
and "watchedprogressbar") anywhere on the screen that will graphically
display the the percentage of the program recorded/watched. The
recorded progress bar is only available while MythTV is actively
recording. It also updates the new button list item progress bars with
the watched and recorded percentages.

The "viewscheduled" screen described in the "schedule-ui.xml" theme
file can embed a progressbar named "recordedprogressbar" anywhere on
the screen that will graphically display the the percentage of the
program already recorded. Again, only available while MythTV is
actively recording. It also updates the second new button list item
progress bar with the recorded percent.

Note: MythTV does not currently remember the time of the last write of
video data, so the "percent recorded" timings are based on wall clock
time. A future enhancement could be to remember the time of the last
write of video data and compare that to the start/end times of the
recording.

Fixes #331.