ChannelEditor sorting and presentation
Add secondary sorting fields in the ORDER BY clauses to give a deterministic
channel order for all sorting orders.
Keep the same channel, as identified by service ID and transport ID, selected when
changing sorting order; previously the channel in the same position in the list was
selected after changing the sort order.
Preserve the position of a channel in the list on the screen as much as possible
when the sorting order is changed.