mingw32: Fix missing SendMessageA function.
This is caused by a indirect include of winuser.h, which
unconditionally redefines the string SendMessage (and a whole bunch of
other strings that match windows API function names) by either
appending the letter 'A' or 'W' to them. This is what is screwing up
the compile on the call to gCoreContext->SendMessage().