Tune with SI table standard (dvb, atsc, mpeg) from database
Previously, tuning was done with the SI table standard of the last tuned channel on this input or from the default value.
The default value at startup is "mpeg" and this is overwritten by the first channel tuning which is done at startup.
Since commits 90f98e621a7efdf0bd3faf3905a7055322d3f24a and 0c6fd96275f63b742d823c1ce0a37aca9990b828
the tuning at startup is only done for the parent (real) inputs and not anymore for the child (virtual) inputs.
This means the default value for the child inputs stays at "mpeg".
The first tuning of a DVB channel on a child input is then done with the wrong SI table standard.
This does actually work but it gives problems when doing channel changes in Live TV.
This is now fixed by tuning with the SI table standard from the database.

Refs #452