HTTP Headers are not case sensitive
Some clients send the headers in mixed case and
some send the headers in lowercase.

In the case of the backend web app (written in Angular)
the 'content-length' header was being sent in lowercase
which failed the comparison with 'Content-Length'. This
lead to the body of the request not being parsed and as
a result API calls were missing their parameters and
throwing errors.