EIT updates of programs with starttime in the past
Changing the starttime or the title of a program that is being
recorded will start another recording of the same program.
Therefore EIT updates with a starttime in the past are now modified
to keep the old starttime and program title before updating the database.
Previously, updates with a starttime in the past were discarded
when the starttime was changed, unless the endtime was later.