Card identification in mythtv-setup
Append the displayname to the capture card device name in the "Capture Cards" page.
The same is now also done in the "Input Connections" page so that both pages
are now consistent in how the capture cards are presented.
Previously, in the "Input Connections" page the content of field inputname was used.
This field contains now the modulation system and this does not identify the card.
This change in presentation is especially useful to identify DVB-S/S2 cards
that are connected to multiple LNBs via a DiSEqC switch, where
each LNB is represented by a separate entry in the capturecard table
for the same capture card device.