coverity: Fix "dereference after null check" warning.
The body of the first "if" clause in this function explicitly tests to
see if the m_recorded variable is null, but the body of the second
"if" clause dereferences the variable without checking. If it can be
null in one place, it can be null in the other. Add a null test to the
second clause to ensure the dereference doesn't fail.