coverity: Fix "explicit null dereference" warning in imagemanager.cpp.
Coverity isn't correctly tracking that the "if (m_devices.value(id))"
statement passing means that the result must not be a nullptr, and in
the very next statement it still will not be a nullptr. All it is
seeing is that the second argument to the value function defaults to a
nullptr, so it assumes that the function can always return a nullptr.
This is a false positive in coverity, but its simple enough to rewrite
the code to eliminate the warning.